requirements.txt
IMDbPY==2022.7.9
matplotlib==3.6.2
numpy==1.22.3
pandas==1.5.2
Pillow==9.3.0
Pillow==9.5.0
Requests==2.30.0
scikit_learn==1.2.2
seaborn==0.12.2
torch==1.12.1
torchvision==0.13.1a0
